Finished Scrap (Locations of the Scrap Variables in Description of Change)
Enemies will add 10 to the value of the int Scrap when they are killed
In the HUD is the Nodes that display the Scrap Value on the Screen
In SurviveGameMode (Not SurviveConceptGameMode) is where the Function/Event to Add score is
In BP_Enemy_Base is where it finds out when to add 10 to Scrap. It was added on the the Do Damage event where it makes the enemy die and spawn the explosion effect.